+++ packages/adders/common.ts
@@ -16,7 +16,7 @@ export function addEslintConfigPrettier(content: string): string {
 	let svelteImportName: string;
 	for (const specifier of sveltePluginImport?.specifiers ?? []) {
 		if (specifier.type === 'ImportDefaultSpecifier' && specifier.local?.name) {
-			svelteImportName = specifier.local.name;
+			svelteImportName = specifier.local.name as string;
 		}
 	}