+++ packages/cli/common.ts
@@ -8,9 +8,52 @@ import { COMMANDS, constructCommand, resolveCommand } from 'package-manager-dete
 import type { Argument, HelpConfiguration, Option } from 'commander';
 import type { AdderWithoutExplicitArgs, Precondition } from '@sveltejs/cli-core';
 
+const NO_PREFIX = '--no-';
+let options: readonly Option[] = [];
+
+function getLongFlag(flags: string) {
+	return flags
+		.split(',')
+		.map((f) => f.trim())
+		.find((f) => f.startsWith('--'));
+}
+
 export const helpConfig: HelpConfiguration = {
 	argumentDescription: formatDescription,
-	optionDescription: formatDescription
+	optionDescription: formatDescription,
+	visibleOptions(cmd) {
+		// hack so that we can access existing options in `optionTerm`
+		options = cmd.options;
+
+		const visible = cmd.options.filter((o) => !o.hidden);
+		const show: Option[] = [];
+		// hide any `--no-` flag variants if there's an existing flag of a similar name
+		// e.g. `--types` and `--no-types` will combine into a single `--[no-]types` flag
+		for (const option of visible) {
+			const flag = getLongFlag(option.flags);
+			if (flag?.startsWith(NO_PREFIX)) {
+				const stripped = flag.slice(NO_PREFIX.length);
+				const isNoVariant = visible.some((o) => getLongFlag(o.flags)?.startsWith(`--${stripped}`));
+				if (isNoVariant) continue;
+			}
+			show.push(option);
+		}
+		return show;
+	},
+	optionTerm(option) {
+		const longFlag = getLongFlag(option.flags);
+		const flag = longFlag?.split(' ').at(0);
+		if (!flag || !longFlag) return option.flags;
+
+		// check if there's a `--no-{flag}` variant
+		const noVariant = `--no-${flag.slice(2)}`;
+		const hasVariant = options.some((o) => getLongFlag(o.flags) === noVariant);
+		if (hasVariant) {
+			return `--[no-]${longFlag.slice(2)}`;
+		}
+
+		return option.flags;
+	}
 };
 
 function formatDescription(arg: Option | Argument): string {
@@ -77,6 +120,7 @@ export async function installDependencies(agent: AgentName, cwd: string) {
 	try {
 		const { command, args } = constructCommand(COMMANDS[agent].install, [])!;
 		await exec(command, args, { nodeOptions: { cwd } });
+
 		spinner.stop('Successfully installed dependencies');
 	} catch (error) {
 		spinner.stop('Failed to install dependencies', 2);